+++ CHANGELOG.md
@@ -19,13 +19,16 @@ Date format: (YYYY-MM-DD)
 * Ocelot shopkeepers of cat types have been converted to corresponding cat shopkeepers.
   * You might have to adapt your players' permissions, since cat and ocelot shopkeepers require different permissions now. With this migration players can end up with cat shopkeepers even though they don't have the permission to freshly create those if they wanted.
   * The different cat types are represented by colored leather armor inside the editor.
-* Any settings affecting regular villagers will affect wandering traders as well now. There are no settings for now to handle wandering traders differently.
-  * When preventing regular villagers from spawning, the trader llamas of the wandering trader are prevented from spawning as well.
-  * When hiring a wandering trader, its leashed llama should get unleashed due to the removal of the wandering trader.
+* Added separate settings for disabling trading, spawning and allowing hiring of wandering traders.
+  * When preventing wandering traders from spawning, trader llamas are prevented from spawning as well.
+  * When a wandering trader is hired, its llamas will remain in the world (just like it is the case when the wandering trader dies).
+* Internal: Spigot is no longer calling creature spawn events for entities that get spawned as part of chunk generation. If spawning of regular villagers or wandering traders is disabled, these get removed during the corresponding chunk load event now.
 * Added: Sign shops can now switch between different wood types.
 * Added all new 1.14 mobs to the by default enabled mobs. If you are updating you will have to manually enable them in the config.
 
 Other changes:  
+* Changed: When spawning of villagers or wandering traders is disabled, villagers and wandering traders spawned by other plugins, spawn eggs, mob spawners or due to curing zombie villagers are still allowed to spawn now.
+* Added a separate setting 'disable-zombie-villager-curing' (default: false) that can be used to prevent curing of zombie villagers.
 * Made shop and object type matching more strict. This uses a fixed list of internal aliases now.
 * Enabled various mobs by default now, which previously had some issues but seem to work fine now: Horse, mule, donkey, skeleton horse, zombie horse, llama.
 * Removed the generic 'sub type' editor option in favor of letting each shop object supply a list of editor options. This allows living shopkeepers to provide multiple editor options now.
@@ -51,6 +54,9 @@ Internal:
 * Fixed a minor internal inconsistency with the updating of trades: Trades were updated on the client, but the server was left in the previous state (mostly affected debug messages).
 * Removed special handling of item damage tags from item comparison. Spigot has made some changes that should make this obsolete.
 * Moved common de/serialization and validation code of shop object properties into new Property classes.
+
+New messages:  
+* msg-zombie-villager-curing-disabled
 
 ## v2.6.0 (2019-03-04)
 ### Supported MC versions: 1.13.2